Removed fp16

+++ mango/transcription.py
@@ -26,8 +26,7 @@ class WhisperTranscriptionConfig:
 class WhisperTranscriptor(BaseTranscriptor):
     def __init__(self, config: WhisperTranscriptionConfig):
         self.config = config
-        self.whisper = WhisperForConditionalGeneration.from_pretrained(config.whisper_checkpoint, device_map='auto',
-                                                                       torch_dtype=torch.float16)
+        self.whisper = WhisperForConditionalGeneration.from_pretrained(config.whisper_checkpoint, device_map='auto')
         self.whisper.eval()
         self.processor = WhisperProcessor.from_pretrained(config.processor_checkpoint)
         self.processor.tokenizer = WhisperTokenizerFast.from_pretrained(config.processor_checkpoint)
@@ -36,7 +35,7 @@ class WhisperTranscriptor(BaseTranscriptor):
     def transcribe(self, waveform: np.ndarray, sr: int = 16_000) -> str:
         waveform = torchaudio.transforms.Resample(sr, 16_000)(torch.tensor(waveform, dtype=torch.float))
         features = self.processor(waveform, return_tensors='pt', sampling_rate=16_000)
-        features = features.to(self.whisper.device).half()
+        features = features.to(self.whisper.device)
         forced_decoder_ids = self.processor.get_decoder_prompt_ids(language="uk", task="transcribe")
 
         generated_ids = self.whisper.generate(**features, forced_decoder_ids=forced_decoder_ids)
